refactor: settings ルートを composition に分離
